Create a list of user-defined parameters

The function creates a list of user-defined parameters.

create_parameter_list(no_predictors = 16, dependent_variable = "FOI",
  resample_grid_size = 20, grid_size = 5, no_samples = 1,
  vec_phis_R0_1 = c(1, 1, 0, 0), vec_phis_R0_2 = c(1, 1, 1, 1),
  prop_sympt = c(0.45, 0.85, 0.15, 0.15), prop_hosp = c(0.04, 0.1,
  0.04, 0.04), FOI_grid = seq(0.002, 0.2, 0.002), sf_vals = seq(1, 0.1,
  -0.1), sat_functions_shapes = c(0, 5, 1600000), no_trees = 500,
  min_node_size = 20, all_wgt = 1, pseudoAbs_value = c(FOI = -0.02,
  R0_1 = 0.5, R0_2 = 0.5, R0_3 = 0.5, Z = -0.02), foi_offset = 0.03,
  EM_iter = 10, base_info = c("cell", "latitude", "longitude",
  "population", "ID_0", "ID_1", "ID_2"), extra_params = NULL)

Arguments

no_predictors
number of selected covariates used for model fitting and making predictions.

dependent_variable
character string of the fitted response variable (FOI).

resample_grid_size
resolution of the foi predictions (in km). Default = 20 (1/6 degree).

grid_size
size of the grid used for block bootstrapping.

no_samples
number of bootstrap samples.

vec_phis_R0_1
numeric vector of length = 4 of the relative infectiousness of primary, secondary, tertiary and quaternary dengue infections, when assuming only primary and secondary infections are infectious.

vec_phis_R0_2
numeric vector of length = 4 of the relative infectiousness of primary, secondary, tertiary and quaternary dengue infections, when assuming all four infections are infectious.

prop_sympt
numeric vector of length = 4 of the proportions of primary, secondary, tertiary and quaternary infections which are symptomatic.

prop_hosp
numeric vector of length = 4 of the proportions of primary, secondary, tertiary and quaternary infections requiring hospitalization.

FOI_grid
numeric vector of force of infection values used for mapping FOI to number of infections, cases hopsitalizations and R0.

sf_vals
scaling factor used to model the effect of transmission-reducing interventions.

sat_functions_shapes
parameters of the saturating function used for setting pseudo absences case weights.

no_trees
number of trees of the random forest model passed to ranger.

min_node_size
minimal node size of the random forest model passed to ranger.

all_wgt
numeric value of case weights for all data points.

pseudoAbs_value
numeric value of pseudo absences for different response variables.

foi_offset
numeric value to offset model FOI predictions during fitting with Expectation Maximization.

EM_iter
number of iterations of the Expectation Maximization algorithm

base_info
character string of key ID variable in the 1/6 degree resolution global covariate dataset.

extra_params
list of additional parameters. Default = NULL.