Fit a random forest using an EM algorithm

The function fits a random forest model to FOI data point estimates at 1/6 degree resolution using an Expectation Maximization algorithm

exp_max_algorithm(parms, adm_dataset, pxl_dataset, covariates_names,
  map_col = NULL, RF_obj_path = NULL, RF_obj_name = NULL,
  diagn_tab_path = NULL, diagn_tab_name = NULL, map_path = NULL,
  sct_plt_path = NULL, train_dts_path = NULL, train_dts_name = NULL,
  adm_covariates = NULL)

Arguments

parms
list of user-defined parameters.

adm_dataset
dataframe of the bootstrapped dataset of foi estimates and covariates at admin unit 1 resolution (pre-processed).

pxl_dataset
dataframe of the bootstrapped dataset of foi estimates and covariates at 1/6 degree resolution.

covariates_names
character vector of covariates names.

map_col
map colours. Default = NULL.

RF_obj_path
character string of the directory for saving the random forest object model. Default = NULL.

RF_obj_name
character of name of the random forest model object to save. Default = NULL.

diagn_tab_path
character string of the directory for saving the table of diagnostics from the EM fit. Default = NULL.

diagn_tab_name
character of name of the EM fit table of diagnostics to save. Default = NULL.

map_path
character string of the directory for saving the figure of the prediction map from the EM fit. Default = NULL.

sct_plt_path
character string of the directory for saving the figure of observation vs predictions plot. Default = NULL.

train_dts_path
character string of the directory for saving the training dataset. Default = NULL.

train_dts_name
character of name of the training dataset dataframe to save. Default = NULL.

adm_covariates
dataframe of global covariates at admin unit 1 resolution. Default = NULL.

Value

the random forest model object returned by ranger.